requirements.txt
alembic==1.8.1
apispec==6.0.2
Babel==2.11.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==21.3
passlib==1.7.4
psycopg2==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
SQLAlchemy==1.4.44
urllib3==1.26.12
webargs==8.2.0
Werkzeug==2.2.2
